- Made a number of fixes to the updater/launcher. - Added 'server news' widget to the launcher. - Warning fixes.

+++ trunk/src/client/psengine.cpp
@@ -974,7 +974,7 @@ const char* psEngine::FindCommonString(unsigned int cstr_id)
    return clientDR->GetMsgStrings()->Request(cstr_id);
 }
 
-size_t psEngine::FindCommonStringId(const char *str)
+uint psEngine::FindCommonStringId(const char *str)
 {
     if (!celclient)
         return csInvalidStringID;
